misc updates
