+++ packages/components/src/editing/hooks/hooksUtils.ts
@@ -1,9 +1,40 @@
 import { SyntaxNode, TreeCursor } from "@lezer/common";
+import {
+  ShapeType,
+  makeCanvas,
+  sampleShape,
+  shapeTypes,
+  simpleContext,
+} from "@penrose/core";
+import { ShapeDefinitions } from "../types";
 
 export const extractText = (progText: string, to: number, from: number) => {
   return progText.slice(from, to);
 };
 
+/**
+ * Retrieves defintions for all shapes and writes their properties to a
+ * hashmap object with shapeName as the top level key with the value being
+ * a hashmap of properties that belong to the shape as keys with type as value
+ */
+export const getShapeDefs = (): ShapeDefinitions => {
+  const shapeProps = {} as ShapeDefinitions;
+  const size = 311; // placeholder, this value doesn't matter
+
+  for (const shapeName of shapeTypes) {
+    const shapeSample = sampleShape(
+      shapeName as ShapeType,
+      simpleContext("ShapeProps dummy"),
+      makeCanvas(size, size),
+    );
+
+    shapeProps[shapeName] = Object.fromEntries(
+      Object.entries(shapeSample).map(([key, value]) => [key, value.tag]),
+    );
+  }
+
+  return shapeProps;
+};
 export const goToParentX = (node: SyntaxNode, x: number) => {
   let i = 0;
   let nextParent: SyntaxNode | null = node;
